20161226 - allow for use of search term, action and model filtering of permissions

+++ app/Http/Controllers/PermissionsController.php
@@ -10,8 +10,7 @@ use Illuminate\Support\Facades\Redirect;
 class PermissionsController extends Controller
 {
     //
-    public function __construct()
-    {
+    public function __construct() {
         $this->middleware('auth');
         if ('cli' != php_sapi_name()) {
             $this->authorize('show-admin-menu');
@@ -20,11 +19,50 @@ class PermissionsController extends Controller
         
     }
     
-    public function index()
-    {
-        //
-        $permissions = \montserrat\Permission::orderBy('name')->get();
-        return view('admin.permissions.index',compact('permissions'));   
+    public function index(Request $request) {
+        $this->authorize('show-permission');
+        $actions = array (
+            '' => 'N/A',
+            'create'=>'create',
+            'delete'=>'delete',
+            'edit'=>'edit',
+            'manage'=>'manage',
+            'update'=>'update'        
+        );
+        $models = array (
+            '' => 'N/A',
+            'address'=>'address',
+            'attachment'=>'attachment',
+            'contact'=>'contact',
+            'email'=>'email',
+            'group'=>'group',
+            'note'=>'note',
+            'permission'=>'permission',
+            'phone'=>'phone',
+            'registration'=>'registration',
+            'relationship'=>'relationship',
+            'reservation'=>'reservation',
+            'retreat'=>'retreat',
+            'role'=>'role',
+            'room'=>'room',
+            'touchpoint'=>'touchpoint',
+            'user'=>'user',
+            'website'=>'website'
+    
+        
+        
+        );
+        if (empty($request->input('term'))) {
+            $term = $request->input('action').'-'.$request->input('model');
+        } else {
+            $term = $request->input('term');
+        }
+        if (empty($term)) {
+            $permissions = \montserrat\Permission::orderBy('name')->get();
+        } else {
+            $permissions = \montserrat\Permission::orderBy('name')->where('name','like','%'.$term.'%')->get();
+        }
+        return view('admin.permissions.index',compact('permissions','actions','models'));   
     }
 
     /**
@@ -32,10 +70,8 @@ class PermissionsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
-        //
-        
+    public function create() {
+        $this->authorize('create-permission');
         return view('admin.permissions.create');  
 
     }
@@ -46,18 +82,14 @@ class PermissionsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-        //
-
-    $permission = new \montserrat\Permission;
-    $permission->name= $request->input('name');
-    $permission->display_name= $request->input('display_name');
-    $permission->description = $request->input('description');
-    
-    $permission->save();
-    
-    return Redirect::action('PermissionsController@index');
+    public function store(Request $request) {
+        $this->authorize('create-permission');
+        $permission = new \montserrat\Permission;
+        $permission->name= $request->input('name');
+        $permission->display_name= $request->input('display_name');
+        $permission->description = $request->input('description');
+        $permission->save();
+        return Redirect::back();
     }
     
 
@@ -67,14 +99,11 @@ class PermissionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-        //
+    public function show($id) {
+        $this->authorize('show-permission');
         $permission = \montserrat\Permission::with('roles')->findOrFail($id);
         $roles = \montserrat\Role::pluck('name','id');
-        
-        return view('admin.permissions.show',compact('permission','roles'));//
-
+        return view('admin.permissions.show',compact('permission','roles'));
     }
 
     /**
@@ -83,13 +112,10 @@ class PermissionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
-        //
+    public function edit($id) {
+        $this->authorize('update-permission');
         $permission = \montserrat\Permission::findOrFail($id);
-        
-        return view('admin.permissions.edit',compact('permission'));//
-
+        return view('admin.permissions.edit',compact('permission'));
     }
 
     /**
@@ -99,19 +125,15 @@ class PermissionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
-        //
+    public function update(Request $request, $id) {
+        $this->authorize('update-permission');
         $permission = \montserrat\Permission::findOrFail($request->input('id'));
-    
         $permission->name= $request->input('name');
         $permission->display_name= $request->input('display_name');
         $permission->description= $request->input('description');
-
         $permission->save();
     
-    return Redirect::action('PermissionsController@index');
-
+        return Redirect::action('PermissionsController@index');
     }
 
     /**
@@ -120,19 +142,19 @@ class PermissionsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        //
+    public function destroy($id) {
+        $this->authorize('delete-permission');
         \montserrat\Permission::destroy($id);
        return Redirect::action('PermissionsController@index');
-    
     }
+    
     public function update_roles(Request $request) {
+        $this->authorize('update-permission');
+        $this->authorize('update-role');
         $permission = \montserrat\Permission::findOrFail($request->input('id'));
         $permission->roles()->detach();
         $permission->roles()->sync($request->input('roles'));
     
         return Redirect::action('PermissionsController@index');
-    }
-    
+    }   
 }
